move files to new  globalhotkey file
